% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mva.methods-package.R
\docType{package}
\name{mva.methods-package}
\alias{mva.methods}
\alias{mva.methods-package}
\title{mva.methods: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
